Refatored routes and created controllers
