+++ src/modules/appointments/infra/http/routes/appointments.routes.ts
@@ -1,34 +1,13 @@
 import { Router } from 'express';
-import { parseISO } from 'date-fns';
-import { container } from 'tsyringe';
-
-import CreateAppointmentService from '@modules/appointments/services/CreateAppointmentService';
 
 import ensureAthenticated from '@modules/users/infra/http/middlewares/ensureAuthenticated';
+import AppointmentsController from '@modules/appointments/infra/http/controllers/AppoimentsController';
 
 const appointmentsRouter = Router();
+const appointmentsController = new AppointmentsController();
 
 appointmentsRouter.use(ensureAthenticated);
 
-// appointmentsRouter.get('/', async (req, res) => {
-//   const appointments = await AppointmentsRepository.find();
-
-//   return res.json(appointments);
-// });
-
-appointmentsRouter.post('/', async (req, res) => {
-  const { provider_id, date } = req.body;
-
-  const parsedDate = parseISO(date);
-
-  const createAppointment = container.resolve(CreateAppointmentService);
-
-  const appointment = await createAppointment.execute({
-    date: parsedDate,
-    provider_id,
-  });
-
-  return res.json(appointment);
-});
+appointmentsRouter.post('/', appointmentsController.create);
 
 export default appointmentsRouter;
