+++ src/modules/users/infra/http/controllers/UsersController.ts
@@ -1,0 +1,22 @@
+import { Request, Response } from 'express';
+import { container } from 'tsyringe';
+
+import CreateUserService from '@modules/users/services/CreateUserService';
+
+export default class UsersController {
+  async create(request: Request, response: Response): Promise<Response> {
+    const { name, email, password } = request.body;
+
+    const createUser = container.resolve(CreateUserService);
+
+    const user = await createUser.execute({
+      name,
+      email,
+      password,
+    });
+
+    delete user.password;
+
+    return response.json(user);
+  }
+}
